requirements.txt
langchain
llama-index
haystack

fastapi
uvicorn

requests
python-dotenv
langchain-openai